backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
face-recognition==1.3.0
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.3.0
python-dotenv==1.0.1
slowapi==0.1.9
aiofiles==23.2.1
httpx==0.27.0